% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combs.R
\name{combs}
\alias{combs}
\title{Compute the Number of Combinations of r Items Chosen From n}
\usage{
combs(n, r)
}
\arguments{
\item{n}{The total number of items.}

\item{r}{The number of items to choose from \code{n}.}
}
\value{
An integer value representing the number of combinations of \code{r} items chosen from \code{n}. If invalid input values are given, an error message is returned.
}
\description{
The function calculates the number of ways to choose \code{r} items from a total of \code{n} items without replacement and where the order does not matter.
}
\details{
It uses the formula:
\deqn{C(n, r) = \frac{n!}{r!(n - r)!}}
}
