% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_range.R
\name{data_range}
\alias{data_range}
\title{Range of a Data Set}
\usage{
data_range(x)
}
\arguments{
\item{x}{a vector or values}
}
\value{
The function returns the range of a data set, defined as the minimum subtracted from the maximum.
}
\description{
This function computes the range of a data set \code{x}.
}
