% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/One_Prop_Int.R
\name{One_Prop_Int}
\alias{One_Prop_Int}
\title{One Sample Confidence Interval for a Population Proportion}
\usage{
One_Prop_Int(x, n, alpha = 0.05)
}
\arguments{
\item{x}{The number of individuals of interest in the sample.}

\item{n}{The sample size.}

\item{alpha}{A value between 0 and 1, where \code{1 - alpha} is the confidence level as a decimal. The default is 0.05, representing a 95\% confidence level.}
}
\value{
The function returns, invisibly, a list containing:
\itemize{
\item \code{sprop}: The sample proportion.
\item \code{critval}: The critical value used in calculations.
\item \code{Lbound}: The lower boundary of the confidence interval.
\item \code{Ubound}: The upper boundary of the confidence interval.
}
You can access these values with expressions like \code{res$sprop}, \code{res$critval}, \code{res$Lbound}, and \code{res$Ubound} where \code{res} is the result of the function.
}
\description{
This function calculates the confidence interval for a population proportion using the normal distribution.
}
\details{
The function will print the confidence level, margin of error, critical value, sample proportion, and the confidence interval (lower and upper bounds).
}
