% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/One_Prop_Test.R
\name{One_Prop_Test}
\alias{One_Prop_Test}
\title{One Sample Hypothesis Test for a Population Proportion}
\usage{
One_Prop_Test(x, n, p0, alt = "two")
}
\arguments{
\item{x}{The number of individuals of interest in the sample.}

\item{n}{The sample size.}

\item{p0}{The hypothesized proportion in the null hypothesis.}

\item{alt}{A character string specifying the alternate hypothesis. Choices are "left", "right", or "two". If omitted, the default is "two". The choices "less", "greater", or "two.sided" are also accepted.}
}
\value{
The function returns, invisibly, a list containing:
\itemize{
\item \code{sprop}: The sample proportion.
\item \code{z}: The test statistic z.
\item \code{pvalue}: The p-value corresponding to the test statistic.
}
You can access these values with expressions like \code{res$sprop}, \code{res$z}, and \code{res$pvalue} where \code{res} is the result of the function.
}
\description{
This function performs a one-sample hypothesis test for a population proportion.
}
\details{
The function will print the sample proportion, test statistic z, and the p-value.
}
