% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_bounds.R
\name{outlier_bounds}
\alias{outlier_bounds}
\title{Outlier Bounds using IQR Method}
\usage{
outlier_bounds(x)
}
\arguments{
\item{x}{a vector or values}
}
\value{
The function returns, invisibly, a list containing:
\itemize{
\item \code{Lower.bound}: The lower outlier bound.
\item \code{Upper.bound}: The upper outlier bound.}
You can access these values with expressions like \code{res$Lower.bound} and \code{res$Upper.bound} where \code{res} is the result of the function.
}
\description{
This function computes the lower and upper outlier boundaries using the IQR method.
}
\details{
The lower outlier boundary is defined by \code{Q1 - 1.5(IQR)} and the upper outlier boundary is defined by \code{Q3 + 1.5(IQR)}. The function will print the results.
}
