% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perms.R
\name{perms}
\alias{perms}
\title{Compute the Number of Permutations of r Items Chosen From n}
\usage{
perms(n, r)
}
\arguments{
\item{n}{The total number of items.}

\item{r}{The number of items to choose from \code{n}.}
}
\value{
An integer value representing the number of permutations of \code{r} items chosen from \code{n}. If invalid input values are given, an error message is returned.
}
\description{
The function calculates the number of ways to choose \code{r} items from a total of \code{n} items without replacement and where the order matters.
}
\details{
It uses the formula:
\deqn{P(n, r) = \frac{n!}{(n - r)!}}
}
