% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rank_Sum_Test.R
\name{Rank_Sum_Test}
\alias{Rank_Sum_Test}
\title{Rank Sum Test for Comparing the Medians of Two Populations.}
\usage{
Rank_Sum_Test(sample1, sample2, alpha, alt = "two.sided")
}
\arguments{
\item{sample1}{A numeric vector representing the first sample data.}

\item{sample2}{A numeric vector representing the second sample data.}

\item{alpha}{The significance level for the hypothesis test (between 0 and 1, exclusive).}

\item{alt}{A character string specifying the alternate hypothesis. Choices are "left", "right", "two", "less", "greater", or "two.sided". If omitted, the default is "two.sided".}
}
\description{
This function performs the nonparametric Rank Sum Test for comparing the medians of populations
}
\details{
The function provides detailed output including the null and alternate hypotheses, significance level, test statistic, p-value, and the test result.
}
