% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rel_hist.R
\name{rel_hist}
\alias{rel_hist}
\title{Relative Histogram}
\usage{
rel_hist(x, ...)
}
\arguments{
\item{x}{a vector of values for the relative histogram}

\item{...}{further arguments and graphical parameters available in the \code{hist} function.}
}
\value{
The function returns, invisibly, a list containing:
\itemize{
\item \code{breaks}: A vector containing the cell boundaries.
\item \code{counts}: A vector containing the relative frequencies for each class.
\item \code{mids}: A vector containing the cell midpoints
}
You can access these values with expressions like \code{res$breaks}, \code{res$counts}, and \code{res$mids} where \code{res} is the result of the function.
}
\description{
The function \code{rel_hist} computes a relative histogram of given data values.
}
