% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd.p.R
\name{sd.p}
\alias{sd.p}
\title{Calculate the population standard deviation}
\usage{
sd.p(x)
}
\arguments{
\item{x}{a vector of values}
}
\value{
The population standard deviation of \code{x}.
}
\description{
The function \code{sd.p} calculates the population standard deviation of a data set \code{x}.
}
