% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sign_Test.R
\name{Sign_Test}
\alias{Sign_Test}
\title{Sign Test for Testing a Median}
\usage{
Sign_Test(sample, m0, alpha, alt = "two.sided")
}
\arguments{
\item{sample}{A numeric vector representing the sample data.}

\item{m0}{Hypothesized population median in the null hypothesis.}

\item{alpha}{The significance level for the hypothesis test (between 0 and 1, exclusive). Choices for one-tailed tests include 0.005, 0.01, 0.025, 0.05. Choices for two-tailed tests include 0.01, 0.02, 0.05, 0.10.}

\item{alt}{A character string specifying the alternate hypothesis. Choices are "left", "right", or "two". If omitted, the default is "two". The choices "less", "greater", or "two.sided" are also accepted.}
}
\description{
This function performs a nonparametric one-sample Sign Test for a population median when the population distribution is not necessarily normal. Critical values are included based on specified alpha levels and alternate hypotheses.
}
\details{
The function calculates the test statistic, critical value, and determines whether to reject the null hypothesis. It provides detailed output including the null and alternate hypotheses, significance level, test statistic, critical value, and the test result.
}
