% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T_Interval.R
\name{T_Interval}
\alias{T_Interval}
\title{T Confidence Interval for a Population Mean}
\usage{
T_Interval(xbar, n, s, alpha = 0.05)
}
\arguments{
\item{xbar}{The sample mean.}

\item{n}{The sample size.}

\item{s}{The sample standard deviation.}

\item{alpha}{A value between 0 and 1, where \code{1 - alpha} is the confidence level as a decimal. The default is 0.05, representing a 95\% confidence level.}
}
\value{
The function returns, invisibly, a list containing:
\itemize{
\item \code{critval}: The critical value used in calculations.
\item \code{Lbound}: The lower boundary of the confidence interval.
\item \code{Ubound}: The upper boundary of the confidence interval.
}
You can access these values with expressions like \code{res$critval}, \code{res$Lbound}, and \code{res$Ubound} where \code{res} is the result of the function.

If invalid input values are given, an error message is returned.
}
\description{
This function calculates the confidence interval for a population mean when the population standard deviation is not known. This confidence interval is based on the Student's t-distribution.
}
\details{
The function will print the confidence level, margin of error, number of degrees of freedom, critical value, and the confidence interval (lower and upper bounds).
}
\note{
This function is appropriate when the population standard deviation is not known. If the population standard deviation is known, consider using the  \code{\link{Z_Interval}} function, based on the normal distribution.
}
