% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T_Test.R
\name{T_Test}
\alias{T_Test}
\title{Hypothesis Test for a Population Mean, Standard Deviation Unknown}
\usage{
T_Test(xbar, n, s, mu0, alt = "two")
}
\arguments{
\item{xbar}{The sample mean.}

\item{n}{The sample size.}

\item{s}{The sample standard deviation.}

\item{mu0}{The hypothesized mean in the null hypothesis.}

\item{alt}{A character string specifying the alternate hypothesis. Choices are "left", "right", or "two". If omitted, the default is "two". The choices "less", "greater", or "two.sided" are also accepted.}
}
\value{
The function returns, invisibly, a list containing:
\itemize{
\item \code{t}: The test statistic t.
\item \code{pvalue}: The p-value corresponding to the test statistic.
}
You can access these values with expressions like \code{res$t}, and \code{res$pvalue} where \code{res} is the result of the function.
}
\description{
This function performs a hypothesis test about a population mean when the population standard deviation is not known. P-values are calculated based on the Student's t-distribution.
}
\details{
The function will print the number of degrees of freedom, the test statistic t and the p-value.
}
\note{
This function is appropriate when the population standard deviation is not known. If the population standard deviation is known, consider using the \code{\link{Z_Test}} function, based on the normal distribution.
}
