% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Two_Prop_Int.R
\name{Two_Prop_Int}
\alias{Two_Prop_Int}
\title{Confidence Interval for the Difference Between Two Proportions}
\usage{
Two_Prop_Int(x1, n1, x2, n2, alpha = 0.05)
}
\arguments{
\item{x1}{The number of individuals of interest in Sample 1.}

\item{n1}{The size of Sample 1.}

\item{x2}{The number of individuals of interest in Sample 2.}

\item{n2}{The size of Sample 2.}

\item{alpha}{A value between 0 and 1, where \code{1 - alpha} is the confidence level as a decimal. The default is 0.05, representing a 95\% confidence level.}
}
\value{
The function returns, invisibly, a list containing:
\itemize{
\item \code{sprop1}: Sample 1 proportion.
\item \code{sprop2}: Sample 2 proportion.
\item \code{point_est}: The point estimate for the confidence interval.
\item \code{critval}: The critical value used in calculations.
\item \code{Lbound}: The lower boundary of the confidence interval.
\item \code{Ubound}: The upper boundary of the confidence interval.
}
You can access these values with expressions like \code{res$sprop1}, \code{res$point_est}, and \code{res$Ubound} where \code{res} is the result of the function.
}
\description{
This function calculates the confidence interval for the difference between two population proportions (\code{p_1 - p_2}).
}
\details{
The function will print the confidence level, margin of error, critical value, point estimate, and the confidence interval (lower and upper bounds).
}
