% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Two_Samp_T_Interval.R
\name{Two_Samp_T_Interval}
\alias{Two_Samp_T_Interval}
\title{Confidence Interval for the Difference Between Two Means: Independent Samples}
\usage{
Two_Samp_T_Interval(xbar1, s1, n1, xbar2, s2, n2, alpha = 0.05)
}
\arguments{
\item{xbar1}{The mean of Sample 1.}

\item{s1}{The standard deviation of Sample 1.}

\item{n1}{The size of Sample 1.}

\item{xbar2}{The mean of Sample 2.}

\item{s2}{The standard deviation of Sample 2.}

\item{n2}{The size of Sample 2.}

\item{alpha}{A value between 0 and 1, where \code{1 - alpha} is the confidence level as a decimal. The default is 0.05, representing a 95\% confidence level.}
}
\value{
The function returns, invisibly, a list containing:
\itemize{
\item \code{critval}: The critical value used in calculations.
\item \code{point_est}: The point estimate for the confidence interval.
\item \code{Lbound}: The lower boundary of the confidence interval.
\item \code{Ubound}: The upper boundary of the confidence interval.
}
You can access these values with expressions like \code{res$critval}, \code{res$Lbound}, and \code{res$Ubound} where \code{res} is the result of the function.
}
\description{
This function constructs a confidence interval for the difference between two population means (\code{mu_1 - mu_2}) given two independent samples. This confidence interval is based on the Student's t-distribution.
}
\details{
The function will print the confidence level, margin of error, number of degrees of freedom, critical value, point estimate, and the confidence interval (lower and upper bounds).
}
