% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Two_Samp_T_Test.R
\name{Two_Samp_T_Test}
\alias{Two_Samp_T_Test}
\title{Hypothesis Test for the Difference Between Two Means: Independent Samples}
\usage{
Two_Samp_T_Test(xbar1, s1, n1, xbar2, s2, n2, alt = "two", mu = 0)
}
\arguments{
\item{xbar1}{The mean of Sample 1.}

\item{s1}{The standard deviation of Sample 1.}

\item{n1}{The size of Sample 1.}

\item{xbar2}{The mean of Sample 2.}

\item{s2}{The standard deviation of Sample 2.}

\item{n2}{The size of Sample 2.}

\item{alt}{A required character string specifying the alternate hypothesis. Choices are "left", "right", or "two". The choices "less", "greater", or "two.sided" are also accepted.}

\item{mu0}{The hypothesized mean in the null hypothesis. If omitted, \code{mu0} is assumed to be 0.}
}
\value{
The function returns, invisibly, a list containing:
\itemize{
\item \code{t}: The test statistic t.
\item \code{pvalue}: The p-value corresponding to the test statistic.
}
You can access these values with expressions like \code{res$t}, and \code{res$pvalue} where \code{res} is the result of the function.

If invalid input values are given, an error message is returned.
}
\description{
This function performs a hypothesis test about the difference between two population means (\code{mu_1 - mu_2}) given two independent samples. P-values are calculated based on the Student's t-distribution.
}
\details{
The function will print the number of degrees of freedom, the test statistic t and the p-value for the test about \code{mu_1 - mu_2}.
}
