% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var.p.R
\name{var.p}
\alias{var.p}
\title{Calculate the population variance}
\usage{
var.p(x)
}
\arguments{
\item{x}{a vector of values}
}
\value{
The population variance of \code{x}.
}
\description{
The function \code{var.p} calculates the population variance of a data set \code{x}.
}
