% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Z_Interval.R
\name{Z_Interval}
\alias{Z_Interval}
\title{Z Confidence Interval for a Population Mean}
\usage{
Z_Interval(xbar, n, sigma, alpha = 0.05)
}
\arguments{
\item{xbar}{The sample mean.}

\item{n}{The sample size.}

\item{sigma}{The known population standard deviation.}

\item{alpha}{A value between 0 and 1, where \code{1 - alpha} is the confidence level as a decimal. The default is 0.05, representing a 95\% confidence level.}
}
\value{
The function returns, invisibly, a list containing:
\itemize{
\item \code{critval}: The critical value used in calculations.
\item \code{Lbound}: The lower boundary of the confidence interval.
\item \code{Ubound}: The upper boundary of the confidence interval.
}
You can access these values with expressions like \code{res$critval}, \code{res$Lbound}, and \code{res$Ubound} where \code{res} is the result of the function.

If invalid input values are given, an error message is returned.
}
\description{
This function calculates the confidence interval for a population mean when the population standard deviation is known. This confidence interval is based on the normal distribution, which is appropriate when the population standard deviation is known.
}
\details{
The function will print the confidence level, margin of error, critical value, and the confidence interval (lower and upper bounds).
}
\note{
This function assumes the population standard deviation is known. If the population standard deviation is not known, consider using the \code{\link{T_Interval}} function, based on the Student's t-distribution.
}
