% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rel_hist.R
\name{rel_hist}
\alias{rel_hist}
\title{Relative Histogram}
\usage{
rel_hist(x, ...)
}
\arguments{
\item{x}{a vector of values for the relative histogram}

\item{main}{Defaults main title to "Relative Histogram" \emph{(internal parameter)}}

\item{ylab}{Defaults the label for the y-axis to "Relative Frequencies" \emph{(internal parameter)}}

\item{ybreaks}{Adjusts y-axis breaks. \emph{(internal parameter)}}

\item{...}{further arguments and graphical parameters available in the \code{hist} function.}
}
\description{
The function \code{rel_hist} computes a relative histogram of given data values.
}
